config: fix code to recompile and gofumpt

+++ config/fake.go
@@ -167,6 +167,6 @@ func (s *FakeService) Subscribe(path string, mr MessageReceiver) (subscriptionID
 
 // Scoped creates a new ScopedReader which uses the underlying
 // mocked paths and values.
-func (s *FakeService) Scoped(websiteID, storeID int64) Scoped {
+func (s *FakeService) Scoped(websiteID, storeID uint32) Scoped {
 	return makeScoped(s, websiteID, storeID)
 }
